check model

+++ app/src/main/java/com/cheese/weatherapp/models/Sys.kt
@@ -1,8 +1,8 @@
 package com.cheese.weatherapp.models
 
 data class Sys (
-    val type: Long,
-    val id: Long,
+    val type: Int,
+    val id: Int,
     val country: String,
     val sunrise: Long,
     val sunset: Long
